requirements.txt
Flask==2.1.1
gspread==5.3.0
oauth2client==4.1.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.26.0
pandas==1.4.1
numpy==1.22.3
scikit-learn==1.0.1
tensorflow==2.7.0
scikit-fuzzy==0.4.0
